Fixes requested by Melissa
I just ran a test and I noticed something… three times now when I visited the main page, clicked on “Companies” and clicked on “Sign Me Up” at the bottom, it directed me to the right page, but it was at the middle of the registration page, not the beginning.

Also, I noticed (twice now) when I clicked on the number of people to have on the 8th…. When I click on “)” it jumped to “10”.  I clicked on “0” gain and it stuck.  But I thought it was odd that it happened twice.  The first time I assumed it was  user error.
Melissa